* pitivi/bin.py: Don't call link/unlink functions if atee/vtee do not exist. This removes some warnings. * pitivi/discoverer.py: Change behaviour of discoverer to handle cases where demuxers do not create all pads at once, but rely on emitting 'no-more-pads' to say there are no more coming. Fixes #353859
git-svn-id: svn+ssh://svn.gnome.org/svn/pitivi/trunk@852 d3729300-e425-0410-8a4c-d956edccc248

+++ pitivi/bin.py
@@ -390,14 +390,20 @@ class SmartFileBin(SmartBin):
         # connect to good tee
         self.debug("SmartFileBin's source has a new pad: %s %s" % (pad , pad.get_caps().to_string()))
         if pad.get_caps().to_string().startswith("audio"):
-            pad.link(self.atee.get_pad("sink"))
+            if not self.atee:
+                self.warning("Got new audio pad, but we didn't discover one previously !")
+            else:
+                pad.link(self.atee.get_pad("sink"))
         elif pad.get_caps().to_string().startswith("video"):
-            pad.link(self.vtee.get_pad("sink"))
+            if not self.vtee:
+                self.warning("Got new video pad, but we didn't discover one previously !")
+            else:
+                pad.link(self.vtee.get_pad("sink"))
 
     def _binRemovedDecodedPadCb(self, unused_bin, pad):
-        if pad.get_caps().to_string().startswith("audio"):
+        if pad.get_caps().to_string().startswith("audio") and self.atee:
             pad.unlink(self.atee.get_pad("sink"))
-        elif pad.get_caps().to_string().startswith("video"):
+        elif pad.get_caps().to_string().startswith("video") and self.vtee:
             pad.unlink(self.vtee.get_pad("sink"))
 
     def do_destroy(self):
